start at code cleanup; PH in this branch will not work
